requirements.txt
psycopg2-binary
neo4j
sentence-transformers>=3.3.1
redis
python-dotenv>=1.0.0
GitPython
openai>=1.0.0
aiohttp>=3.8.0
numpy>=2.2.0
prompt_toolkit>=3.0.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
PyYAML>=6.0.1
jenkinsapi>=0.3.13
jproperties>=2.1.1
javaproperties>=0.8.0
torch>=2.5.1
transformers>=4.47.1
scikit-learn>=1.6.0
networkx>=3.4.2
